refactor: extract MethodDisplayHelpers from GetOverloadsLogic
