+++ src/RoslynCodeLens/Tools/GetOverloadsLogic.cs
@@ -1,5 +1,3 @@
-using System.Globalization;
-using System.Xml;
 using Microsoft.CodeAnalysis;
 using RoslynCodeLens.Models;
 using RoslynCodeLens.Symbols;
@@ -99,8 +97,8 @@ public static class GetOverloadsLogic
         return new OverloadInfo(
             Signature: method.ToDisplayString(SignatureFormat),
             ReturnType: method.ReturnType.ToDisplayString(),
-            Parameters: method.Parameters.Select(BuildParameter).ToList(),
-            Accessibility: AccessibilityToString(method.DeclaredAccessibility),
+            Parameters: method.Parameters.Select(MethodDisplayHelpers.BuildParameter).ToList(),
+            Accessibility: MethodDisplayHelpers.AccessibilityToString(method.DeclaredAccessibility),
             IsStatic: method.IsStatic,
             IsVirtual: method.IsVirtual,
             IsAbstract: method.IsAbstract,
@@ -108,90 +106,8 @@ public static class GetOverloadsLogic
             IsAsync: method.IsAsync,
             IsExtensionMethod: method.IsExtensionMethod,
             TypeParameters: method.TypeParameters.Select(t => t.Name).ToList(),
-            XmlDocSummary: ExtractSummary(method),
+            XmlDocSummary: MethodDisplayHelpers.ExtractSummary(method),
             FilePath: file,
             Line: line);
-    }
-
-    private static OverloadParameter BuildParameter(IParameterSymbol p)
-    {
-        var defaultText = p.HasExplicitDefaultValue
-            ? FormatDefault(p.ExplicitDefaultValue, p.Type)
-            : null;
-
-        var modifier = p.RefKind switch
-        {
-            RefKind.Ref => "ref",
-            RefKind.Out => "out",
-            RefKind.In => "in",
-            RefKind.RefReadOnlyParameter => "ref readonly",
-            _ => string.Empty,
-        };
-
-        return new OverloadParameter(
-            Name: p.Name,
-            Type: p.Type.ToDisplayString(),
-            IsOptional: p.IsOptional,
-            DefaultValue: defaultText,
-            IsParams: p.IsParams,
-            Modifier: modifier);
-    }
-
-    private static string FormatDefault(object? value, ITypeSymbol type)
-    {
-        // Enum defaults: ExplicitDefaultValue returns the underlying integer (e.g. 0). Look
-        // up the matching field on the enum so we render `MyEnum.Foo` instead of `0`.
-        if (value is not null && type.TypeKind == TypeKind.Enum && type is INamedTypeSymbol enumType)
-        {
-            foreach (var member in enumType.GetMembers().OfType<IFieldSymbol>())
-            {
-                if (member.HasConstantValue && Equals(member.ConstantValue, value))
-                    return $"{enumType.Name}.{member.Name}";
-            }
-        }
-
-        return value switch
-        {
-            null => "null",
-            string s => $"\"{s}\"",
-            bool b => b ? "true" : "false",
-            char c => $"'{c}'",
-            IFormattable f => f.ToString(null, CultureInfo.InvariantCulture),
-            _ => value.ToString() ?? "null",
-        };
-    }
-
-    private static string AccessibilityToString(Accessibility a) => a switch
-    {
-        Accessibility.Public => "public",
-        Accessibility.Internal => "internal",
-        Accessibility.Protected => "protected",
-        Accessibility.ProtectedAndInternal => "private protected",
-        Accessibility.ProtectedOrInternal => "protected internal",
-        Accessibility.Private => "private",
-        _ => "internal",
-    };
-
-    private static string? ExtractSummary(IMethodSymbol method)
-    {
-        var xml = method.GetDocumentationCommentXml();
-        if (string.IsNullOrWhiteSpace(xml)) return null;
-
-        try
-        {
-            var doc = new XmlDocument();
-            doc.LoadXml(xml);
-            var summary = doc.SelectSingleNode("//summary");
-            var text = summary?.InnerText.Trim();
-            if (string.IsNullOrEmpty(text)) return null;
-
-            // InnerText flattens nested tags (e.g. <see cref="X"/>) to empty strings, leaving
-            // double-spaces. Collapse whitespace runs to single spaces for clean rendering.
-            return System.Text.RegularExpressions.Regex.Replace(text, @"\s+", " ");
-        }
-        catch (XmlException)
-        {
-            return null;
-        }
     }
 }
